keep schema.ts separate
