+++ packages/react-sdk/src/css/index.ts
@@ -1,6 +1,7 @@
 export * from "./properties";
 export * from "./units";
 export * from "./types";
+export * from "./schema";
 export * from "./get-browser-style";
 export * from "./categories";
 export * from "./keyword-values";
