+++ packages/react-sdk/src/user-props/index.ts
@@ -1,3 +1,4 @@
 export * from "./all-user-props";
 export * from "./use-user-props";
 export * from "./types";
+export * from "./schema";
