+++ packages/react-sdk/src/user-props/types.ts
@@ -1,16 +1,5 @@
-import { z } from "zod";
 import type { InstanceProps, Instance } from "../db";
-
-export const UserProp = z.object({
-  id: z.string(),
-  prop: z.string(),
-  value: z.union([z.string(), z.boolean()]),
-  required: z.optional(z.boolean()),
-});
-
-export const UserProps = z.array(UserProp);
-
-export type UserProp = z.infer<typeof UserProp>;
+import { UserProp } from "./schema";
 
 export type UserPropsUpdates = {
   treeId: InstanceProps["treeId"];
